requirements.txt
Django==1.11.21
pytz==2017.2
psycopg2==2.7.3.1
python-dotenv==0.9.1
environs==7.1.0